requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.7
Django==5.0.6
django-allauth==0.62.1
django-dotenv==1.4.2
django-storages==1.14.3
django-timezone-field==6.1.0
djangorestframework==3.15.1
gcloud==0.17.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
jws==0.1.3
kombu==5.3.7
lxml==5.2.2
msgpack==1.0.8
mysqlclient==2.2.4
oauth2client==3.0.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.4.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
